python-app/requirements.txt
Flask==2.3.3
python-consul2==0.1.5
